requirements.txt
kafka-python
mysql-connector-python
flask
flask-cors
python-dotenv
certifi
cors